refactor(client): extract @kolu/platform browser/OS adapters
Move keyboard (chord matching), os (Apple-modifier detection), and clipboard
(non-secure-context-safe writes + xterm provider) into a reusable package.
These are framework-agnostic — no solid-js dependency — so the package omits
the solid- prefix the other extractions carry. 23 import sites rewritten via
subpath exports; keyboard.test stays in client (it pairs keyboard with Kolu
ACTIONS/prohibitedKeybinds) and now consumes the package.

Behaviour-preserving. client/src: 24,405 -> 24,201 LOC (-204).
