+++ packages/client/src/CommandPalette.tsx
@@ -24,7 +24,7 @@ import {
 } from "solid-js";
 import { Dynamic } from "solid-js/web";
 import { match } from "ts-pattern";
-import { formatKeybind, type Keybind } from "./input/keyboard";
+import { formatKeybind, type Keybind } from "@kolu/platform/keyboard";
 import { useTips } from "./settings/useTips";
 import Kbd from "@kolu/solid-ui/Kbd";
 import ModalDialog from "./ui/ModalDialog";
