+++ packages/client/src/MobileChromeSheet.tsx
@@ -18,7 +18,7 @@
 
 import { type Component, createSignal } from "solid-js";
 import { ACTIONS } from "./input/actions";
-import { formatKeybind } from "./input/keyboard";
+import { formatKeybind } from "@kolu/platform/keyboard";
 import { useRightPanel } from "./right-panel/useRightPanel";
 import type { WsStatus } from "./rpc/rpc";
 import SettingsPopover from "./settings/SettingsPopover";
