+++ packages/client/src/SplitStrip.tsx
@@ -3,7 +3,7 @@
 
 import type { Component } from "solid-js";
 import { ACTIONS } from "./input/actions";
-import { formatKeybind } from "./input/keyboard";
+import { formatKeybind } from "@kolu/platform/keyboard";
 import Kbd from "@kolu/solid-ui/Kbd";
 
 type SplitStripProps =
