+++ packages/client/src/input/actions.ts
@@ -14,7 +14,7 @@ import type { TerminalId, TerminalMetadata } from "kolu-common/surface";
 import { nonEmpty } from "nonempty";
 import type { Accessor, Setter } from "solid-js";
 import type { PaletteAction, SectionId } from "../CommandPalette";
-import { type Keybind, matchesKeybind } from "./keyboard";
+import { type Keybind, matchesKeybind } from "@kolu/platform/keyboard";
 
 /** Shared handler context — every dispatched action receives this. */
 export interface ActionContext {
