+++ packages/client/src/input/zoom.ts
@@ -10,7 +10,7 @@ import { makePersisted } from "@solid-primitives/storage";
 import { DEFAULT_FONT_SIZE } from "kolu-common/config";
 import type { TerminalId } from "kolu-common/surface";
 import { createSignal } from "solid-js";
-import { isPlatformModifier, ZOOM_KEYS } from "./keyboard";
+import { isPlatformModifier, ZOOM_KEYS } from "@kolu/platform/keyboard";
 
 /**
  * Reactive font-size signal driven by Cmd/Ctrl +/- and Cmd/Ctrl+0 (reset).
