+++ packages/platform/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@kolu/platform",
+  "version": "0.1.0",
+  "private": true,
+  "license": "AGPL-3.0-or-later",
+  "type": "module",
+  "sideEffects": false,
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./keyboard": "./src/keyboard.ts",
+    "./clipboard": "./src/clipboard.ts",
+    "./os": "./src/os.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "@types/node": "^22.0.0",
+    "typescript": "^5.8.0"
+  },
+  "dependencies": {
+    "@xterm/addon-clipboard": "^0.2.0"
+  }
+}
