+++ packages/platform/src/clipboard.ts
